Extract FlowController state snapshot
Committed-By-Agent: codex
